🎨 Move send pose and transpose pose to own module
